pyproject.toml
[project]
name = "tikal-dbt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click==8.1.8",
    "click-command-tree==1.2.0",
    "click-completion==0.5.2",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "clickloader==0.2.0",
    "cookiecutter==2.6.0",
    "dbt-artifacts-parser==0.8.2",
    "dbt-bigquery>=1.9.1",
    "dbt-core==1.9.3",
    "dotenv==0.9.9",
    "elementary-data[bigquery]==0.16.2",
    "pyyaml==6.0.2",
    "rich-click==1.8.6",
    "singleton-decorator>=1.0.0",
    "snowflake-connector-python==3.13.2",
    "snowflake-snowpark-python>=1.28.0",

]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest-cov>=6.0.0",
    "sqlfluff>=3.3.1",
    "sqlfluff-templater-dbt>=3.3.1",
    "tox>=4.24.2",
]

[tool.setuptools.packages.find]
include = ["cli*"]  

[project.scripts]
cli = "cli.dbt_cli:start"
cli_validate = "cli.validate.validate_hook:main"
